Revert Event back (from RefCountedWillBeRefCountedGarbageCollected) to RefCounted
Moving Event to oilpan's heap caused a lot of crashes/failures. So this CL reverts Event back to RefCounted.

This CL is a revert for https://codereview.chromium.org/182063004/. However, this CL does not revert trace() methods in the original CL because there is no harm in having trace() methods in the code base. (If I revert trace() methods, I have to resolve a bunch of conflicts manually, which will be more dangerous than leaving trace() methods in the code base.)

BUG=340522
NOTRY=true

Review URL: https://codereview.chromium.org/185463005

git-svn-id: svn://svn.chromium.org/blink/trunk@168354 bbb929c8-8fbe-4397-9dbb-9b2b20218538