Remove CanvasRenderingContext from the web exposed objects
We used have a shared super class called CanvasRenderingContext for
CanvasRenderingContext2D and WebGLRenderingContext. This does not
match the spec and it does not match Firefox or IE.

This removes the web exposed idl file for CanvasRenderingContext. The
method that used to return a CanvasRenderingContext now return a
(CanvasRenderingContext2D or WebGLRenderingContext) instead.

This also allows us to remove the custom bindings that was needed for
CanvasRenderingContext

This is a web exposed visible change. CanvasRenderingContext was
discoverable by walking the prototype chain. It was never exposed as a
named property of Window.

BUG=348012
R=ch.dumez@samsung.com, kbr@chromium.org, tkent@chromium.org

Committed: https://src.chromium.org/viewvc/blink?view=rev&revision=2cda866

Review URL: https://codereview.chromium.org/177983005

git-svn-id: svn://svn.chromium.org/blink/trunk@168488 bbb929c8-8fbe-4397-9dbb-9b2b20218538