Maintain SkPaint in GraphicsContextState.
The aim is to avoid SkPaint setup time for the common case of filling and stroking shapes. We can't entirely do that, as some of the paint content is not controlled by GraphicsContextState (such as text or temporary paint state set for specific draw operations).

Perf tests on the record_and_rasterize micro-bench indicate a 1-3% average improvement on record times.

BUG=343240
R=adamk@chromium.org, fmalita@chromium.org, senorblanco@chromium.org

Review URL: https://codereview.chromium.org/169283008

git-svn-id: svn://svn.chromium.org/blink/trunk@168536 bbb929c8-8fbe-4397-9dbb-9b2b20218538