Revert "Don't precompile scripts." (r168112)
This reverts revision 168112.

Reason: the morejs.warm_times regressed by 2-7%. The benchmark is a bit
artificial wrt renderer reuse, but I believe it captures an important
aspect that no other benchmark captures: Navigating inside the same site
will also reuse renderers, so in-memory caching in renderers might be
a good thing. (Note that the preparse data is never written to disk and
is not shared between renderers.)

BUG=348311
TBR=marja@chromium.org,haraken@chromium.org

Review URL: https://codereview.chromium.org/190473003

git-svn-id: svn://svn.chromium.org/blink/trunk@168726 bbb929c8-8fbe-4397-9dbb-9b2b20218538