Don't account for box-shadow in the outline box
The outline box is supposed to be taking care of the outline as we
don't account for it in the visual overflow. box-shadow is already
factored into the visual overflow through
RenderBox::addVisualEffectOverflow.

The 3 differences in the tests are invalidations that corresponds
to the intersection between the old or new box and the box shadow.
The tests still paint correctly and the invalidations make sense
but could be the indication that we don't invalidate correctly
(part of the box shadow is not invalidated but those cases are
correct by skipping it).

BUG=345452

Review URL: https://codereview.chromium.org/189163008

git-svn-id: svn://svn.chromium.org/blink/trunk@168774 bbb929c8-8fbe-4397-9dbb-9b2b20218538