Use outlineBox if we have an outline
Currently we decided to repaint if the outline box has moved or if we have to
paint the background and our outline box has changed shape. This CL updates the
code to only look at the outlineBox if we have an outline.

The patch was largely done by Dan Sinclair (dsinclair@chromium.org) on
https://codereview.chromium.org/187813004/. I merely generated the test
expectations and added some code to avoid over-invalidating.

BUG=320139

Review URL: https://codereview.chromium.org/192863004

git-svn-id: svn://svn.chromium.org/blink/trunk@168968 bbb929c8-8fbe-4397-9dbb-9b2b20218538