Revert XMLHttpRequest's receive-as-blob implementation back to use memory until Chromium is fixed
The RedirectToFile functionality of Chromium on which the current
optimized [1] receive-as-blob code of XHR depends is broken.
Because of the bug, XHR::didDownloadData() handler may be invoked after
XHR::didFinishLoading(). See [2] for detail. As it seems the fix is not
ready for landing but the impact of the bug is big, I'd like to first
revert the optimization in Blink until Chromium is fixed well.

[1] http://src.chromium.org/viewvc/blink?view=revision&revision=163141
[2] https://codereview.chromium.org/82273002/

BUG=347663

Review URL: https://codereview.chromium.org/190583004

git-svn-id: svn://svn.chromium.org/blink/trunk@168993 bbb929c8-8fbe-4397-9dbb-9b2b20218538