Remove obsolete FIXME that no longer applies for blink rendering architecture
After studying this FIXME and the motivation behind it I attempted to fix it by
clipping to the viewport as suggested.  However, it turns out this FIXME is no
longer applicable because the architecture of Blink's rendering routine requires
all Blink to inform the compositor of all invalidations not just the visible ones.

Review URL: https://codereview.chromium.org/184043033

git-svn-id: svn://svn.chromium.org/blink/trunk@169065 bbb929c8-8fbe-4397-9dbb-9b2b20218538