CompositedLayerMapping's dirty bits don't account for tree changes
When removing a CompositedLayerMapping, we correctly marked it as dirty.
However, the m_needToUpdateGeometryOfAllDecendants bit was being lost because
we'd remove the CompositedLayerMapping from the tree and the descendants
wouldn't know they needed a geometry update.

In principle, we could mark all of our descendants as needing a geometry update,
but finding all the descendants requires walking the z-order lists, which can
require a lot of work. Instead, this CL over invalidates by marking our parent
as needing a geometry update. That will force all our children to be updated as
well.

BUG=350854

Review URL: https://codereview.chromium.org/196383003

git-svn-id: svn://svn.chromium.org/blink/trunk@169070 bbb929c8-8fbe-4397-9dbb-9b2b20218538