Oilpan: Allow allocations during finalization.
This is the simplest possible patch to allow allocation during
finalization. I would like to land it to unblock work that is
currently waiting for the ability to allocate during finalization.

Once this lands I want to iterate on it to make sure that:

1. We can allocate from memory that has been found to be free
during the current collection.

2. Make sure that the pages allocated for allocations during
finalization are cleanly separated from the rest of the pages.

R=erik.corry@gmail.com, vegorov@chromium.org, oilpan-reviews@chromium.org

Review URL: https://codereview.chromium.org/197603003

git-svn-id: svn://svn.chromium.org/blink/trunk@169118 bbb929c8-8fbe-4397-9dbb-9b2b20218538