WTF::Vector::swap() should not mem{cpy,move} unsupported classes.
Prior to this CL, swap was always done by memcpy when the vector size was within its inlineCapacity.
However, this caused problems when the class was not memcpy-able.

This CL adds VectorTraits<T>::canSwapWithMemcpy, and VectorSwapper template.
Only classes which are marked canSwapWithMemcpy are swapped with memcpy, and all other classes are swapped via copy c-tor and operator=.

BUG=267349
R=ager@chromium.org, erik.corry@gmail.com, jochen@chromium.org

Review URL: https://codereview.chromium.org/194363002

git-svn-id: svn://svn.chromium.org/blink/trunk@169217 bbb929c8-8fbe-4397-9dbb-9b2b20218538