Web Animations API: Load resources referenced in element.animate()
This patch fixes a bug where resources in element.animate() calls were
not being fetched. This resulted in StylePendingImages leaking out of
the style resolving stage and into rendering which causes a crash.

This patch fixes the bug by ensuring StylePendingImages introduced in
applyAnimatedProperties() will be resolved as non-pending StyleImages
through an additional call to loadPendingResources().
This patch also modifies AnimatableImage to be a wrapper of CSSValues
instead of StyleImages in order to follow the same resource caching
logic as typical CSS image values.
This patch also updates EffectInput::convert() to pass the document's
StyleSheetContents to MutableStylePropertySet::setProperty() so that
relative URLs are correctly resolved to absolute URLs.

BUG=353385

Review URL: https://codereview.chromium.org/196573030

git-svn-id: svn://svn.chromium.org/blink/trunk@169424 bbb929c8-8fbe-4397-9dbb-9b2b20218538